added confirm option for deleting an user

+++ src/components/Users/Users.js
@@ -10,18 +10,22 @@ const Users = () => {
     }, []);
 
     const handleDelete = id => {
-        const url = `http://localhost:5000/users/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-        .then(res => res.json())
-        .then(data => {
-            if(data.deletedCount > 0){
-                alert("deleted successfully!!");
-                const remainingUsers = users.filter(user => user._id !== id);
-                setUsers(remainingUsers);
-            }
-        })
+        const proceed = window.confirm("Are you sure, you want to delete??");
+
+        if(proceed){
+            const url = `http://localhost:5000/users/${id}`;
+            fetch(url, {
+                method: 'DELETE'
+            })
+            .then(res => res.json())
+            .then(data => {
+                if(data.deletedCount > 0){
+                    alert("deleted successfully!!");
+                    const remainingUsers = users.filter(user => user._id !== id);
+                    setUsers(remainingUsers);
+                }
+            });
+        }
     }
 
     return (
